crates/sel4-hal-adapters: Rename module
