+++ crates/sel4-hal-adapters/src/lib.rs
@@ -7,4 +7,4 @@
 
 #![no_std]
 
-pub mod embedded_hal;
+pub mod serial;
