+++ crates/sel4-hal-adapters/src/serial/mod.rs
